+++ packages/UsageAnalysis/files/TestTrack/Viewers/BoxPlot/box-plot-spec.ts
@@ -1,6 +1,6 @@
 import {test, expect} from '@playwright/test';
-import {loginToDatagrok, specTestOptions, softStep, login} from '../spec-login';
-import * as v from '../helpers/viewers';
+import {loginToDatagrok, specTestOptions, softStep, login} from '../../spec-login';
+import * as v from '../../helpers/viewers';
 
 test.use(specTestOptions);